Revert "added drivers/gsl from coda-oss:main"
This reverts commit 19a9c7445c72518506c5201a2d00227d8241c585.

+++ externals/coda-oss/modules/c++/sys/include/sys/ReadWriteMutex.h
@@ -44,7 +44,7 @@ class ReadWriteMutex
 {
     public:
     //!  Constructor
-    ReadWriteMutex(int maxReaders) : mSem(static_cast<unsigned int>(maxReaders))
+    ReadWriteMutex(int maxReaders) : mSem(maxReaders)
     {
         mMaxReaders = maxReaders;
         dbg_printf("Creating a read/write mutex\n");
